requirements.txt
h5py==2.10.0
imutils==0.5.4
numba==0.46.0
numpy==1.19.5
Pillow==7.2.0
pipenv==2018.11.26
pipreqs==0.4.10
scikit-image==0.16.2
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-addons==0.12.1
tensorflow-datasets==1.0.2
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
tensorflow-hub==0.11.0
tensorflow-metadata==0.13.0
tensorflow-model-optimization==0.5.0
tensorflow-object-detection-api==0.1.1
tensorflow-probability==0.11.0
tensorflow-tensorboard==0.4.0rc3
tensorflow-text==2.4.3
tf-models-official==2.4.0
tf-slim==1.1.0
tfa-nightly==0.13.0.dev20210324143012
